Add files via upload
This project uses the K-means clustering algorithm to segment customers based on shopping patterns. The dataset contains various attributes, such as InvoiceDate, UnitPrice, and CustomerID. The goal is to help the supermarket identify unsatisfied customer needs, develop appealing products and services, and improve its marketing and distribution strategies. This project provides an opportunity to learn about customer segmentation and its advantages in real-world scenarios. 

This project is designed to solve the problem of understanding customer behaviour and preferences in a supermarket store. By analysing essential customer data such as age, gender, annual income, and spending score, the goal is to identify target customers and provide insights to the marketing team to plan effective marketing strategies. This can be useful for potential users, such as supermarket owners and marketers, who want to improve customer engagement, increase sales, and enhance customer satisfaction.

For this, we analysed the data set of an online store where we grouped the store's customers into different clusters based on their purchasing habits. The different steps involved were:
Missing values treatment
Data transformation
Outlier treatment
Data standardisation
Finding the optimal value of K
Implementing the K-means algorithm
Analysing the clusters of customers to obtain business insights

Once we finish the data preparation, the K-means algorithm becomes easy to implement. All it takes is running the KMeans() function. We only need to decide the number of required clusters beforehand and run the algorithm multiple times with a different number of K before we can figure out the most optimal number of groups.